pyproject.toml
[tool.poetry]
name = "SK-Bot"
version = "0.1.0"
description = "一些乱七八糟的NoneBot2插件"
authors = ["SK-415 <2967923486@qq.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-alpha.13"
nb-cli = "^0.4.1"
Pillow = "^8.0.1"
pillow = "^8.2.0"
nonebot-plugin-apscheduler = "^0.1.2"
nonebot-adapter-cqhttp = "^2.0.0-alpha.12"
haruka-bot = "^1.2.3"

[tool.poetry.dev-dependencies]

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
